Finish dda implementation

+++ include/Line.h
@@ -12,14 +12,14 @@ namespace tiara{
 
 	class Line : public Shape{
 		private:
-			size_t x1, x2, y1, y2;
+			int x1, x2, y1, y2;
 			Color color;
 			LineImpl lineImpl;
 
 			void dda(Canvas & canvas);
 		public:
 
-			Line(size_t x1_, size_t y1_, size_t x2_, size_t y2_, const Color & color_, const tiara::LineImpl lineImpl_ = LineDDA);
+			Line(int x1_, int y1_, int x2_, int y2_, const Color & color_, const tiara::LineImpl lineImpl_ = LineDDA);
 			void draw(Canvas & canvas) override;
 	};
